+++ app/Config/Routes.php
@@ -48,6 +48,8 @@ $routes->group('/', ['filter' => 'login'], function($routes){
 		$routes->get('list', 'CDR\Users::list');
 		$routes->get('status', 'CDR\Users::doStatus');
 		$routes->get('delete', 'CDR\Users::delete');
+		$routes->get('profile/(:num)', 'CDR\Users::profile/$1');
+		$routes->post('profile', 'CDR\Users::store_profile');
 	});
 });